% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Showing where accidents occured}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{the}{state number}

\item{the}{considered year}
}
\value{
a graphical object
}
\description{
This function a graphic showing in a map where the accident occured given a state number and a year
}
\details{
If the state number is wrong (>56) an error message is printed.
If the state considered contains no accident, a message is printed saying
there is no accident to plot
}
\examples{

MyMap<- fars_map_state(1,2013)

}

