% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the csv.bz2 file which the data are read from.}
}
\value{
a data frame tbl
}
\description{
This function reads a FARS dataset containig in csv file given in argument. If the file doesn't exist
the function stops and prints an arror message saying that the data doesn't exist.
If the data exist, the functions reads it and create a data frame from it.
}
\details{
If the format of the file is not csv.bz2 or csv, the function returns an error message saying "invalid file argument"
}
\examples{

acc2013 <- fars_read ("accident_2013.csv.bz2")

}

