% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creating a list of several FARS dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{a}{list of years}
}
\value{
a list of data frame tbl corresponding to the list of years entered in parameters
}
\description{
This function creates a list of truncated FARS datatsets giving th considered years into a list.
Returned FARS datasets are truncated because they only conatines the variables year and month
}
\details{
if a unknown year is given into the list of year, an error message is printed
indicating that the year is invalid. And so the corresponding element of the returned list is NULL
}
\examples{

dat_list <- fars_read_years(list(2013,2014,2015))

}

