% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing the number of events by year and month in a tbl_df object}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{a}{list of years}
}
\value{
a tbl_object corresponding to the list of years entered in parameters
and containing the number of events for each month of each year
}
\description{
This function creates a tabl_df object containing the numbers of events by year and month
contained in the different FARS datasets
}
\details{
if a unknown year is given into the list of year, an error message is printed
indicating that the year is invalid. And so the corresponding element of return table is empty
}
\examples{

MySummary<- fars_summarize_years(list(2013,2015))

}

