% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Returning a a FARS filename according a year entered as a parameter}
\usage{
make_filename(year)
}
\arguments{
\item{a}{Year}
}
\value{
a character vector
}
\description{
This function return a character vector containing a formatted combination of text and year value
}
\details{
the file names created is in csv.bz2 format
}
\examples{

name_2017 <- make_filename(2017)
name_2017
[1] "accident_2017.csv.bz2"

}

